# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizApp API Documentation",
        "contact": {
            "email": "haykalcs@gmail.com"
        },
        "version": "1"
    },
    "paths": {
...
        "/api/materi": {
            "get": {
                "tags": [
                    "Materi"
                ],
                "summary": "Get List of Materi",
                "description": "Get List of Materi",
                "operationId": "getMateriList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "subject": {
                                                        "type": "string"
                                                    },
                                                    "competency": {
                                                        "type": "string"
                                                    },
                                                    "class": {
                                                        "type": "string"
                                                    },
                                                    "semester": {
                                                        "type": "string"
                                                    },
                                                    "meet": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "user_id": 2,
                                                    "subject": "IPA",
                                                    "competence": "Mendeskripsikan tata surya",
                                                    "class": "VI",
                                                    "semester": "2",
                                                    "meet": "1",
                                                    "description": "Penjelasan materi tata surya",
                                                    "image_banner": "1830621834.jpg",
                                                    "created_at": "2021-09-26T12:56:48.000000Z",
                                                    "updated_at": "2021-09-26T13:00:38.000000Z",
                                                    "user": {
                                                        "id": 2,
                                                        "name": "guru",
                                                        "email": "guru@gmail.com",
                                                        "avatar": null
                                                    }
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/materi/{id}": {
            "get": {
                "tags": [
                    "Materi"
                ],
                "summary": "Get Materi Detail",
                "description": "Get Materi Detail",
                "operationId": "showMateri",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Materi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail data"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "user_id": 2,
                                                "subject": "IPA",
                                                "competence": "Mendeskripsikan tata surya",
                                                "class": "VI",
                                                "semester": "2",
                                                "meet": "1",
                                                "description": "Penjelasan materi tata surya",
                                                "image_banner": "1830621834.jpg",
                                                "created_at": "2021-09-26T12:56:48.000000Z",
                                                "updated_at": "2021-09-26T13:00:38.000000Z",
                                                "user": {
                                                    "id": 2,
                                                    "name": "guru",
                                                    "email": "guru@gmail.com",
                                                    "avatar": null
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/guru/materi/{id}/image": {
            "delete": {
                "tags": [
                    "Materi"
                ],
                "summary": "Delete Materi Image",
                "description": "Delete Materi Image",
                "operationId": "deleteImageMateri",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Materi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gambar berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Image not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}